root/apps/workers/scout/requirements.txt
faststream[rabbit,cli]
pydantic-settings
selectolax
curl_cffi
httpx
redis
